Self Review: Refacto provider logic in a hook for better usage
