+++ src/context/FontContext.tsx
@@ -1,7 +1,5 @@
 'use client';
 import {createContext, ReactNode, useEffect, useState} from 'react';
-import {loadFont} from '@remotion/google-fonts/NotoSans';
-import {loadGoogleFont} from '../app/utils/loadFont';
 
 interface FontContextType {
 	selectedFont: string;
@@ -20,28 +18,8 @@ interface FontProviderProps {
 export const FontProvider: React.FC<FontProviderProps> = ({children}) => {
 	const [selectedFont, setSelectedFont] = useState<string>('');
 
-	useEffect(() => {
-		const storedFont = sessionStorage.getItem('selectedFont');
-		if (storedFont) {
-			setSelectedFont(storedFont);
-			loadGoogleFont(storedFont);
-		} else {
-			setSelectedFont('Noto Sans');
-			loadFont();
-		}
-	}, [selectedFont]);
-
-	const handleSetSelectedFont = (font: string) => {
-		setSelectedFont(font);
-		if (font) {
-			sessionStorage.setItem('selectedFont', font);
-		}
-	};
-
 	return (
-		<FontContext.Provider
-			value={{selectedFont, setSelectedFont: handleSetSelectedFont}}
-		>
+		<FontContext.Provider value={{selectedFont, setSelectedFont}}>
 			{children}
 		</FontContext.Provider>
 	);
